extracted out scouting entry type
