+++ src/lib/teamStatsTypes.ts
@@ -1,63 +1,5 @@
-export interface ScoutingEntry {
-  matchNumber: string;
-  alliance: string;
-  scouterInitials: string;
-  selectTeam: string;
-  eventName: string;
-  startPoses0: boolean;
-  startPoses1: boolean;
-  startPoses2: boolean;
-  startPoses3: boolean;
-  startPoses4: boolean;
-  startPoses5: boolean;
-  autoCoralPlaceL1Count: number;
-  autoCoralPlaceL2Count: number;
-  autoCoralPlaceL3Count: number;
-  autoCoralPlaceL4Count: number;
-  autoCoralPlaceDropMissCount: number;
-
-  autoCoralPickPreloadCount: number;
-  autoCoralPickStationCount: number;
-  autoCoralPickMark1Count: number;
-  autoCoralPickMark2Count: number;
-  autoCoralPickMark3Count: number;
-
-  autoAlgaePlaceNetShot: number;
-  autoAlgaePlaceProcessor: number;
-  autoAlgaePlaceDropMiss: number;
-  autoAlgaePlaceRemove: number;
-
-  autoAlgaePickReefCount: number;
-  autoAlgaePickMark1Count: number;
-  autoAlgaePickMark2Count: number;
-  autoAlgaePickMark3Count: number;
-
-  autoPassedStartLine: boolean;
-
-  teleopCoralPlaceL1Count: number;
-  teleopCoralPlaceL2Count: number;
-  teleopCoralPlaceL3Count: number;
-  teleopCoralPlaceL4Count: number;
-  teleopCoralPlaceDropMissCount: number;
-  teleopCoralPickStationCount: number;
-  teleopCoralPickCarpetCount: number;
-
-  teleopAlgaePlaceNetShot: number;
-  teleopAlgaePlaceProcessor: number;
-  teleopAlgaePlaceDropMiss: number;
-  teleopAlgaePlaceRemove: number;
-  teleopAlgaePickReefCount: number;
-  teleopAlgaePickCarpetCount: number;
-
-  shallowClimbAttempted: boolean;
-  deepClimbAttempted: boolean;
-  parkAttempted: boolean;
-  climbFailed: boolean;
-
-  playedDefense: boolean;
-  brokeDown: boolean;
-  comment: string;
-}
+// Import centralized ScoutingEntry type
+export type { ScoutingEntry } from './scoutingTypes';
 
 export interface TeamStats {
   matchesPlayed: number;
